+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * The "Combine Properties" optimization spell can now be used when processing multiple NIF files.
 * Implemented "Flip Normals" for BSTriShape geometry.
+* New spell (Sanitize/Sort Keys) for sorting animation key groups by time. It is applicable to arrays with a type of Key or QuatKey, and a size of at least 2.
 * Fixed glTF import of skinned models where the bone nodes are after the meshes.
 * Fixed rendering Skyrim decals.
 * Fixed issue https://github.com/gavrant/nifskope/issues/8.